Separate the functions that parse the user input
Move the functions that parse the file given as input from the user to a
new file.
