+++ hops/package.go
@@ -15,7 +15,6 @@
 package hops
 
 import (
-	"bytes"
 	"context"
 	"encoding/base64"
 	"encoding/json"
@@ -24,10 +23,7 @@ import (
 	"strings"
 
 	"github.com/moby/buildkit/client/llb"
-	"github.com/moby/buildkit/frontend/dockerfile/instructions"
-	"github.com/moby/buildkit/frontend/dockerfile/parser"
 	ocispecs "github.com/opencontainers/image-spec/specs-go/v1"
-	"gopkg.in/yaml.v3"
 )
 
 const (
@@ -210,125 +206,6 @@ func ToPack(h Hops, buildContext string) (*PackInstructions, error) {
 	instr.Base = rootfsState
 
 	return instr, nil
-}
-
-// ParseBunnyFile reads a yaml file which contains instructions for
-// bunny.
-func ParseBunnyFile(fileBytes []byte, buildContext string) (*PackInstructions, error) {
-	var bunnyHops Hops
-
-	err := yaml.Unmarshal(fileBytes, &bunnyHops)
-	if err != nil {
-		return nil, err
-	}
-
-	err = CheckBunnyfileVersion(bunnyHops.Version)
-	if err != nil {
-		return nil, err
-	}
-
-	err = ValidatePlatform(bunnyHops.Platform)
-	if err != nil {
-		return nil, err
-	}
-
-	err = ValidateKernel(bunnyHops.Kernel)
-	if err != nil {
-		return nil, err
-	}
-
-	// Set default value of from to scratch
-	// Make sure that any reference to Rootfs.From can not be an empty string
-	if bunnyHops.Rootfs.From == "" {
-		bunnyHops.Rootfs.From = "scratch"
-	}
-	err = ValidateRootfs(bunnyHops.Rootfs)
-	if err != nil {
-		return nil, err
-	}
-
-	return ToPack(bunnyHops, buildContext)
-}
-
-// ParseDockerFile reads a Dockerfile-like file and returns a Hops
-// struct with the info from the file
-func ParseDockerFile(fileBytes []byte, buildContext string) (*PackInstructions, error) {
-	instr := new(PackInstructions)
-	instr.Annots = make(map[string]string)
-	instr.Base = llb.Scratch()
-	BaseString := ""
-
-	r := bytes.NewReader(fileBytes)
-
-	// Parse the Dockerfile
-	parseRes, err := parser.Parse(r)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to parse data as Dockerfile: %v", err)
-	}
-
-	// Traverse Dockerfile commands
-	for _, child := range parseRes.AST.Children {
-		cmd, err := instructions.ParseInstruction(child)
-		if err != nil {
-			return nil, fmt.Errorf("Line %d: %v", child.StartLine, err)
-		}
-		switch c := cmd.(type) {
-		case *instructions.Stage:
-			// Handle FROM
-			if BaseString != "" {
-				return nil, fmt.Errorf("Multi-stage builds are not supported")
-			}
-			BaseString = c.BaseName
-		case *instructions.CopyCommand:
-			// Handle COPY
-			var aCopy PackCopies
-
-			aCopy.SrcState = llb.Local(buildContext)
-			aCopy.SrcPath = c.SourcePaths[0]
-			aCopy.DstPath = c.DestPath
-			instr.Copies = append(instr.Copies, aCopy)
-		case *instructions.LabelCommand:
-			// Handle LABEL annotations
-			for _, kvp := range c.Labels {
-				annotKey := strings.Trim(kvp.Key, "\"")
-				instr.Annots[annotKey] = strings.Trim(kvp.Value, "\"")
-			}
-		case instructions.Command:
-			// Catch all other commands
-			return nil, fmt.Errorf("Unsupported command: %s", c.Name())
-		default:
-			return nil, fmt.Errorf("Not a command type: %s", c)
-		}
-
-	}
-	instr.Base = getBase(BaseString, instr.Annots["com.urunc.unikernel.hypervisor"])
-
-	return instr, nil
-}
-
-// ParseFile identifies the format of the given file and either calls
-// ParseDockerFile or ParseBunnyFile
-func ParseFile(fileBytes []byte, buildContext string) (*PackInstructions, error) {
-	lines := bytes.Split(fileBytes, []byte("\n"))
-
-	// First line is always the #syntax
-	if len(lines) <= 1 {
-		return nil, fmt.Errorf("Invalid format of file")
-	}
-
-	// Simply check if the first non-empty line starts with FROM
-	// If it starts we assume a Dockerfile
-	// otherwise a bunnyfile
-	for _, line := range lines[1:] {
-		if len(bytes.TrimSpace(line)) > 0 {
-			if strings.HasPrefix(string(line), "FROM") {
-				return ParseDockerFile(fileBytes, buildContext)
-			}
-			break
-		}
-	}
-
-	return ParseBunnyFile(fileBytes, buildContext)
 }
 
 // Create a LLB State that simply copies all the files in the include list inside
